morph/requirements.txt
# MorphLLM Integration Requirements
requests>=2.31.0
aiohttp>=3.8.0
pydantic>=2.0.0
kubernetes>=27.2.0
pyyaml>=6.0.0
python-dotenv>=1.0.0  # For loading environment variables

# Optional: For enhanced functionality
openai>=1.0.0  # If using OpenAI models with MorphLLM
anthropic>=0.7.0  # If using Claude models with MorphLLM
